JPA file refactored, dead code deleted
